Refactor: reduce and simplify data downloads code
Also remove rows when they are null or the date is not valid.

This means Autograph is definitely useless for downloading anything but timeseries data. Which is okay because there are loose assumptions all over the codebase that could causes issues if you tried to use it for non-timeseries data. We'll hopefully do categorical data properely later.